requirements.txt
fastapi
uvicorn
jinja2
pyyaml
requests
docker 